Shuffle: fix instances of 32-bit int used instead of 64-bit (#1906)
Note that the type used by std::accumulate to compute the sum is
based on the `init` value which if simply 0 means it will use
int, regardless of type of input vectors or output.

+++ bodo/io/_csv_json_reader.cpp
@@ -1183,7 +1183,7 @@ void balance_rows(MemReader *reader) {
 
     // get total number of rows in global dataset
     int64_t total_rows =
-        std::accumulate(num_rows_ranks.begin(), num_rows_ranks.end(), 0);
+        std::accumulate(num_rows_ranks.begin(), num_rows_ranks.end(), int64_t(0));
 
     // by default don't send or receive anything. this is changed below as
     // needed
